[rename] 자동 오프라인 처리 서비스 파일명 변경 #230
