+++ frontend/lib/service/auto_offline_service.dart
